戻り値にファイル名が入った配列が入ってしまう問題を修正、ファイル名の変更
