--- 05.ls/aaa.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-type = 'test'
-lists = {
-  'interview' => 'インタビュー',
-  'test' => 'テスト',
-  'standard' => '通常'
-}
-p lists[type]
