--- 05.ls/iiiiiiiii.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-iiiiiiiii.rb
