+++ 05.ls/sample.txt
@@ -1,3 +1,3 @@
 # frozen_string_literal: true
 
-aaaaaaaaaaa.rb
+ls.rb
